Add useEvents hook
